requirements.txt
web3>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
eth-account>=0.9.0
asyncio>=3.4.3
py-clob-client>=0.21.0
